Moving QA function to validation checks
