+++ mbs_results/utilities/validation_checks.py
@@ -1,7 +1,10 @@
+import logging
 import warnings
 from importlib import metadata
 
 import pandas as pd
+
+logger = logging.getLogger(__name__)
 
 
 def validate_config(config):
@@ -256,3 +259,88 @@ def validate_outlier_detection(df: pd.DataFrame, config: dict):
     snapshot_name = config["mbs_file_name"].split(".")[0]
     outlier_filename = f"outlier_output_v{file_version_mbs}_{snapshot_name}.csv"
     df.to_csv(output_path + outlier_filename)
+
+
+def qa_selective_editing_outputs(config: dict):
+    """
+    function to QA check the selective editing outputs
+    Outputs warnings to logging file if any issues are found
+
+    Parameters
+    ----------
+    config : dict
+        main config for pipeline
+    """
+
+    # Loading SE outputs, function to create SE outputs cannot return them, easier to
+    # load them here
+
+    logger.info("QA checking selective editing outputs")
+
+    file_version_mbs = metadata.metadata("monthly-business-survey-results")["version"]
+    snapshot_name = config["mbs_file_name"].split(".")[0]
+    se_contributor_path = (
+        config["output_path"]
+        + f"selective_editing_contributor_v{file_version_mbs}_{snapshot_name}.csv"
+    )
+    se_question_path = (
+        config["output_path"]
+        + f"selective_editing_question_v{file_version_mbs}_{snapshot_name}.csv"
+    )
+
+    contributor_df = pd.read_csv(se_contributor_path).rename(
+        columns={"ruref": "reference"}
+    )
+    question_df = pd.read_csv(se_question_path).rename(columns={"ruref": "reference"})
+
+    # Checking that references match
+    contributor_unique_reference = contributor_df["reference"].unique().tolist()
+    question_unique_reference = question_df["reference"].unique().tolist()
+    unmatched_references = list(
+        set(contributor_unique_reference).symmetric_difference(
+            set(question_unique_reference)
+        )
+    )
+
+    if len(unmatched_references) > 0:
+        logger.warning(
+            f"There are {len(unmatched_references)} unmatched references in the"
+            " contributor and question SE outputs"
+            f"unmatched references {unmatched_references}"
+        )
+
+    # Checking for duplicates
+    groupby_cols = {
+        "contributor": ["period", "reference"],
+        "question": ["period", "reference", "question_code"],
+    }
+    dataframe_dict = {"contributor": contributor_df, "question": question_df}
+    for dataframe_name in ["contributor", "question"]:
+        dataframe = dataframe_dict.get(dataframe_name)
+        duplicated = dataframe[
+            dataframe.duplicated(subset=groupby_cols[dataframe_name], keep=False)
+        ]
+        if duplicated.shape[0] > 0:
+            logger.warning(
+                f"""There are {duplicated.shape[0]}
+            duplicated {dataframe_name} in the SE outputs"""
+            )
+        else:
+            logger.info(
+                f"no duplicates in {dataframe_name} dataframe columns "
+                f"{groupby_cols[dataframe_name]}"
+            )
+
+        if dataframe.isnull().sum(axis=0).any():
+            null_columns = dataframe.isnull().sum(axis=0)
+            null_columns = null_columns[null_columns > 0]
+            if not null_columns.empty:
+                logger.warning(
+                    f"Nulls or NaNs detected in se {dataframe_name} "
+                    "dataframe in the following columns:\n"
+                    f"{null_columns}"
+                )
+        else:
+            logger.info(f"No nulls or NaNs detected in {dataframe_name} dataframe")
+
+    logger.info("QA of SE outputs finished")
